requirements.txt
glcontext==3.0.0
glfw==2.9.0
moderngl==5.12.0
numpy==2.2.6
pillow==11.2.1
pyglm==2.8.2
